Load config and dynamically import app bundle
Jira ticket: CAMS-389
