+++ .secrets.baseline
@@ -118,9 +118,9 @@
         "filename": "user-interface/index.html",
         "hashed_secret": "c73412a0985c449a8ef92a680bdf8bd0eb829839",
         "is_verified": false,
-        "line_number": 22
+        "line_number": 21
       }
     ]
   },
-  "generated_at": "2025-05-07T23:05:11Z"
+  "generated_at": "2025-05-09T17:16:57Z"
 }
