+++ user-interface/index.html
@@ -5,7 +5,6 @@
     <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-    <link rel="apple-touch-icon" href="/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
@@ -23,6 +22,6 @@
       crossorigin="anonymous"
     ></script>
     <script src="https://atl.cld.prd.ust.doj.gov/jira/plugins/servlet/issueCollectorBootstrap.js?collectorId=4ee4702d&locale=en_US"></script>
-    <script type="module" src="/src/index.tsx"></script>
+    <script type="module" src="src/initialize.ts"></script>
   </body>
 </html>
